Fix geo printing in info command
The geo printing was failing for anything that had two parts.
The curl results we get is a json file and unfortunately awk inteprets
every single space as a "new field". Thus any information that had
two parts in like, North Ostrobothnia would come out only as "North".

We already use jq in other places in the same script, so why not use
that instead? With -r option we do not even need any funky sed things
to strip parenthesis etc.