fw-tools/testnet.sh - fix comments, check k8/gw
Fix comments from top (no need to remain in same folder anymore).
Check also now the k8s and gateways connection, as the edge cases
will require check - we should verify that connection to those
works as well.

Removed the snap connectivity, as that is only applicable to
snap (and it has it's tooling in its own repo).